+++ templates/content-site/template/cms/server/cms/articles/articles.go
@@ -1,7 +1,0 @@
-package articles
-
-import "github.com/gosimple/slug"
-
-func GenerateSlug(name string) string {
-	return slug.Make(name)
-}