+++ templates/content-site/template/cms/server/cms/test_helpers/client/multipart.go
@@ -1,7 +1,0 @@
-package client
-
-type MulipartRequest struct {
-	FileField string
-	FilePath  string
-	Content   []byte
-}