+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## master
+* Remove unused dir in content site template
+
 ## 6.1.1
 * Fix session name in admin panel template
 * Fix `goose_test_up` command in admin panel & content site tempaltes